feat: Room Entity 생성
