+++ src/main/java/com/beyond/easycheck/EasycheckApplication.java
@@ -8,8 +8,6 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @EnableJpaAuditing
 @EnableTransactionManagement
 @SpringBootApplication
-@EnableJpaAuditing
-@EnableTransactionManagement
 public class EasycheckApplication {
 
     public static void main(String[] args) {
