+++ src/main/java/com/beyond/easycheck/room/infrastructure/persistence/entity/RateType.java
@@ -1,0 +1,5 @@
+package com.beyond.easycheck.room.infrastructure.persistence.entity;
+
+public enum RateType {
+    주중, 주말
+}
