+++ src/main/java/com/beyond/easycheck/room/infrastructure/persistence/entity/RoomStatus.java
@@ -1,0 +1,5 @@
+package com.beyond.easycheck.room.infrastructure.persistence.entity;
+
+public enum RoomStatus {
+    예약가능, 예약불가
+}
