+++ src/main/java/com/beyond/easycheck/room/infrastructure/persistence/repository/RoomRateRepository.java
@@ -1,0 +1,7 @@
+package com.beyond.easycheck.room.infrastructure.persistence.repository;
+
+import com.beyond.easycheck.room.infrastructure.persistence.entity.RoomRateEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoomRateRepository extends JpaRepository<RoomRateEntity, Long> {
+}
