+++ src/main/java/com/beyond/easycheck/room/infrastructure/persistence/repository/RoomRepository.java
@@ -1,0 +1,7 @@
+package com.beyond.easycheck.room.infrastructure.persistence.repository;
+
+import com.beyond.easycheck.room.infrastructure.persistence.entity.RoomEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoomRepository extends JpaRepository<RoomEntity, Long> {
+}
