+++ src/main/java/com/beyond/easycheck/room/ui/requestbody/Room/RoomCreateRequest.java
@@ -1,6 +1,6 @@
-package com.beyond.easycheck.rooms.ui.requestbody;
+package com.beyond.easycheck.room.ui.requestbody.Room;
 
-import com.beyond.easycheck.rooms.infrastructure.persistence.entity.RoomStatus;
+import com.beyond.easycheck.room.infrastructure.persistence.entity.RoomStatus;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
