+++ src/main/java/com/beyond/easycheck/room/ui/requestbody/Season/SeasonCreateRequest.java
@@ -1,0 +1,10 @@
+package com.beyond.easycheck.room.ui.requestbody.Season;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SeasonCreateRequest {
+
+}
