--- src/main/java/com/beyond/easycheck/rooms/infrastructure/persistence/entity/RateType.java
@@ -1,5 +1,0 @@
-package com.beyond.easycheck.rooms.infrastructure.persistence.entity;
-
-public enum RateType {
-    주중, 주말
-}
