--- src/main/java/com/beyond/easycheck/rooms/infrastructure/persistence/entity/RoomStatus.java
@@ -1,5 +1,0 @@
-package com.beyond.easycheck.rooms.infrastructure.persistence.entity;
-
-public enum RoomStatus {
-    예약가능, 예약불가
-}
